pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "monarch-mcp-server"
version = "0.1.0"
description = "Model Context Protocol (MCP) server for Monarch Money personal finance platform"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "Rob Cerda" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.0.0",
    "monarchmoney>=0.1.15",
    "gql>=3.4,<4.0",
    "keyring>=24.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
local = [
    "playwright>=1.40.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]

[project.urls]
Homepage = "https://github.com/robcerda/monarch-mcp-server"
Repository = "https://github.com/robcerda/monarch-mcp-server"
Issues = "https://github.com/robcerda/monarch-mcp-server/issues"

[project.scripts]
monarch-mcp-server = "monarch_mcp_server.server:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]